bootstrap: Remove empty lines; refer to bootstrap-dotfiles.sh script
There are many changes in this commit. It makes the Bash script more
modern, and removes some things that I don't use anymore:

1. I moved from Vim to Emacs. So, I don't need YouCompleteMe and other
packages which tried to convert Vim into an IDE. I use it only as a
shell editor now
2. I moved all package installation from apt-get into a single command.
3. I removed the installation of rails and ruby. I don't use ruby
daily, anymore. And this step takes too long and takes up too much space
on the machine. So, it is not required
4. Remove empty lines between comment and commands, and improve comments
to always start with uppercase letter, and be concise
5. Include configuration to install prezto properly. We were not
updating the submodules for prezto, which causes issues on a brand new
installation
6. Include configuration to install Tmux Plugin Manager. This step was
manual before
7. Remove all "yes/no" prompts and replace them with an informative
message that is printed at the end of the script. The user can read the
message and run whatever commands are necessary