zsh: wireguard: Use awk instead of gawk
There is only the GNU flavor of Awk in Ubuntu machines. On Mac OS, awk generally refers to the BSD
awk and GAWK refers to the GNU Awk. Wireguard is a tool which I will use only on my personal
devices, and those will always be Linux devices. So, I can use =awk= here.