zsh: fzf: Update the way in which fzf keybindings and completion is activated
I did not know how this worked until now. This information is there in the README of the fzf package
on GitHub: https://github.com/junegunn/fzf#using-linux-package-managers

I don't know what the content of the =.fzf.zsh= file should be. A copy of that file is not present
anywhere in this repository. So, I want to stop using it.