zsh: Not setting up HTTPS anymore for IP address
It is not required really. Public IP address of a user is known to the ISP anyway. When using a VPN,
the destination IP address of packets seen by the ISP is the same as the public IP address. Using
this on Tor is not a good idea, as the ISP will find your public Tor IP address and be able to
connect that address with you.