emacs: Simplify Vim configuration significantly
I am not using Vim on a day-to-day basis anymore. I still use it sometimes when I am editing commit
messages or making some minor changes. The presence of Vundle and the other "complex" package
managers in the vimrc leads to some problems because they require a local version of Python installed.